+++ src/components/Asset/AssetActions/Pool/Graph/index.module.css
@@ -1,5 +1,5 @@
 .graphWrap {
-  min-height: 97px;
+  min-height: 90px;
   display: flex;
   align-items: center;
   justify-content: center;
